+++ libs/react-client/src/state.ts
@@ -4,6 +4,7 @@ import { Socket } from 'socket.io-client';
 import { v4 as uuidv4 } from 'uuid';
 
 import { ICommand } from './types/command';
+import { IMode } from './types/mode';
 
 import {
   IAction,
@@ -72,6 +73,11 @@ export const messagesState = atom<IStep[]>({
 
 export const commandsState = atom<ICommand[]>({
   key: 'Commands',
+  default: []
+});
+
+export const modesState = atom<IMode[]>({
+  key: 'Modes',
   default: []
 });
 
@@ -236,29 +242,29 @@ export const currentThreadIdState = atom<string | undefined>({
 
 const localStorageEffect =
   <T>(key: string): AtomEffect<T> =>
-  ({ setSelf, onSet }) => {
-    // When the atom is first initialized, try to get its value from localStorage
-    const savedValue = localStorage.getItem(key);
-    if (savedValue != null) {
-      try {
-        setSelf(JSON.parse(savedValue));
-      } catch (error) {
-        console.error(
-          `Error parsing localStorage value for key "${key}":`,
-          error
-        );
+    ({ setSelf, onSet }) => {
+      // When the atom is first initialized, try to get its value from localStorage
+      const savedValue = localStorage.getItem(key);
+      if (savedValue != null) {
+        try {
+          setSelf(JSON.parse(savedValue));
+        } catch (error) {
+          console.error(
+            `Error parsing localStorage value for key "${key}":`,
+            error
+          );
+        }
       }
-    }
 
-    // Subscribe to state changes and update localStorage
-    onSet((newValue, _, isReset) => {
-      if (isReset) {
-        localStorage.removeItem(key);
-      } else {
-        localStorage.setItem(key, JSON.stringify(newValue));
-      }
-    });
-  };
+      // Subscribe to state changes and update localStorage
+      onSet((newValue, _, isReset) => {
+        if (isReset) {
+          localStorage.removeItem(key);
+        } else {
+          localStorage.setItem(key, JSON.stringify(newValue));
+        }
+      });
+    };
 
 export const mcpState = atom<IMcp[]>({
   key: 'Mcp',