+++ libs/react-client/src/types/config.ts
@@ -66,6 +66,7 @@ export interface IChainlitConfig {
     audio: IAudioConfig;
     unsafe_allow_html?: boolean;
     user_message_autoscroll?: boolean;
+    assistant_message_autoscroll?: boolean;
     latex?: boolean;
     user_message_markdown?: boolean;
     edit_message?: boolean;
@@ -90,5 +91,6 @@ export interface IChainlitConfig {
   threadSharing?: boolean;
   chatProfiles: ChatProfile[];
   starters?: IStarter[];
+
   translation: object;
 }